+++ java/client/src/test/java/glide/api/models/StandaloneTransactionTests.java
@@ -1,16 +1,16 @@
 /** Copyright GLIDE-for-Redis Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api.models;
 
+import static glide.api.models.TransactionTests.buildArgs;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static redis_request.RedisRequestOuterClass.RequestType.Move;
 import static redis_request.RedisRequestOuterClass.RequestType.Select;
 
-import com.google.protobuf.ByteString;
 import java.util.LinkedList;
 import java.util.List;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.Test;
 import redis_request.RedisRequestOuterClass;
-import redis_request.RedisRequestOuterClass.Command.ArgsArray;
 
 public class StandaloneTransactionTests {
     @Test
@@ -20,8 +20,9 @@ public class StandaloneTransactionTests {
         Transaction transaction = new Transaction();
 
         transaction.select(5L);
-        results.add(
-                Pair.of(Select, ArgsArray.newBuilder().addArgs(ByteString.copyFromUtf8("5")).build()));
+        results.add(Pair.of(Select, buildArgs("5")));
+        transaction.move("testKey", 2L);
+        results.add(Pair.of(Move, buildArgs("testKey", "2")));
 
         var protobufTransaction = transaction.getProtobufTransaction().build();